refactor: change auth/presentation file structure
