+++ androidApp/src/main/java/com/example/tasky/android/auth/di/LoginModule.kt
@@ -1,6 +1,6 @@
 package com.example.tasky.android.auth.di
 
-import com.example.tasky.android.auth.viewmodel.LoginViewModel
+import com.example.tasky.android.auth.presentation.login.screen.LoginViewModel
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
